+++ example/test/counter/counter_test.dart
@@ -14,31 +14,26 @@ void main() {
   });
   group('Counter', () {
     testWidgets('Increment', (tester) async {
+      //Scenario: Increment
       reporter.startScenario('Increment');
       final examples = [
-        {'value': '1', 'expectedvalue': '1'},
-        {'value': '2', 'expectedvalue': '2'},
-        {'value': '3', 'expectedvalue': '3'},
+        {'value': '1','expectedvalue': '1',},
+        {'value': '2','expectedvalue': '2',},
+        {'value': '3','expectedvalue': '3',},
       ];
       for (var example in examples) {
-        await reporter.guard(
-          () => IncrementScenario.iHaveACounterWithValue0(tester),
-          'Given I have a counter with value 0',
-        );
-        await reporter.guard(
-          () => IncrementScenario.iIncrementTheCounterBy(
-            tester,
-            example['value']!,
-          ),
-          'When I increment the counter by <value>',
-        );
-        await reporter.guard(
-          () => IncrementScenario.theCounterShouldHaveValue(
-            tester,
-            example['expected_value']!,
-          ),
-          'Then the counter should have value <expected_value>',
-        );
+      await reporter.guard(
+        () => IncrementScenario.iHaveACounterWithValue0(tester), 
+        'Given I have a counter with value 0',
+      );
+      await reporter.guard(
+        () => IncrementScenario.iIncrementTheCounterBy(tester,  example['value']!), 
+        'When I increment the counter by <value>',
+      );
+      await reporter.guard(
+        () => IncrementScenario.theCounterShouldHaveValue(tester,  example['expected_value']!), 
+        'Then the counter should have value <expected_value>',
+      );
       }
     });
   });