+++ lib/src/feature/builder/domain/validators/decorators_validator.dart
@@ -8,6 +8,7 @@ class DecoratorsValidator {
     DecoratorType.disableReporter,
     DecoratorType.unitTest,
     DecoratorType.widgetTest,
+    DecoratorType.ignore,
   ];
 
   static final _validScenarioDecoratorTypes = [